Stop reading bulk data segment in file header reads
This was removed long time ago on feature-bulk-data-parsing, as it's not guaranteed that the data was complete, and it's actually useless without a BulkDataHeader, with already has an offset, payload size, and flags. Payloads can be also inline, i.e. appear immediately after the data header. Offset field in the header is rarely used (controlled by flag 0x0001).