Make hierarchy cache not depend on game version
We now have to fetch the game version from the public server network's API, which means it is no longer ties to the Steam data installed locally. Although this is acceptable for JSON version fields, it is not acceptable for the hierarchy cache.

This change removes the public game version from the hierarchy cache's key, leaving only the Steam buildid.