Switch linting from pylint to flake8
Pylint was generating an annoying number of incorrect suggestions due to its limited ability to parse certain forms of code, in particular in mods.

Flake8 is simpler and faster. In combination with mypy it should cover most of our needs with a lot less annoyance.

* Linter in VSCode config switched over
* pylint is not removed yet - just disabled

It is a goal to make the codebase flake8-clean now, in addition to mypy-clean as it currently is.