Fix extensive lint issues for PyRight
These changes are required as part of testing PyLance/PyRight.

Issues are in about equal numbers of positives and negatives. Some cases are due to better handling of types, higlighting places that could not previously be checked, while other cases are due to what look like immaturity issues in PyRight.