Add exporting of alt stats for species
These alt stats are the so-called Troodonisms, where Ark incorrectly applies DCSC priorities in some cases, resulting in incorrect stat values.

Although we have not narrowed down exactly when this happens, and do not know a reliable way to correct it within Ark, we can match the calculation by applying an alternative priority scheme during DCSC data gathering.

With this change we now additionally export any stats that differ from the normal/correct stats when utilising this alternative extraction method.