Update overrides.yaml
The Marnii Wildlife colour regions are an absolute mess so I went through and determined what the regions actually referred to. All Marnii animals possess all 5 colour regions, but when a new creature is spawned, some species don't have any colour in their region (Soul Ball readout says, for example, [3:N] for no colour in region 3). Some species, like the Shadowcat, spawn in with no colour in a region but if a command is used to force a colour into that region, it DOES show up visibly. Other creatures have colours in all five regions, but not all those regions visibly show up on the animal. In many cases this seems as though it was unintended and may be fixed at a later date, so I have differentiated between no colour in a region (null # None, no color generated in colour region), and colour in a region that does not show on the animal (null # Not visible) to make future updates a little easier.