Reformat server version hook and control opendir calls
Replaced tabs with spaces and fixed bracket spacing across the file to remain consistent (with other parts of file and editorconfig dotfile).

opendir(const char*) calls are now intercepted. With the hook activated, the game will be unable to acquire a valid handle to engine localization data and game content directories - which are scanned repeatedly (for unknown reasons) when files are being opened.
The changes do not seem to have any impact on the server in its limited lifetime, and can proceed enough into the initialization process to print the game version. 
This has yielded considerable improvements in tests on WSL 2, and was additionally checked by coldino in a Docker for Windows container (both Hyper-V with the game data residing on a host's filesystem). Previously, Purlovia terminated the process for running for too long, and used an alternative source.
The changes also seem to have a positive impact on non-virtualized Linux systems with direct filesystem access.