Add amazon.ion to deps
Don't forget to run `pipenv sync -d --keep-outdated`.

Even though this is only required for an experimental script, we can't avoid added the dependency due to mypy. Luckily the package is tiny.

Also locked mypy to v0.812. Work up date to v0.900+ will have to be done separately as it will be quite involved.