Add CLI options for extracting one-time mods
This is a much faster way of extracting data for single mods, without the overhead of editing config and the associated hierarchy rediscovery.

The `--mods` CLI option has been extended to include a `+` modifier, which allows mods to be included which are not present in config.

Example:
```
python -m automate asb --mods=+2603240080
```

Note: Mods handled in this way are not uninstalled.