Tentatively fix parsing for Aquatica
This allows asset parsing for the all-important properties but not the extra data such as world tiles and bulk data. NOT YET TESTED ON PREAQUATICA ASSETS.

Note that full scheduled extraction to Obelisk will not continue for a while as the game data itself is still very messed up - with a ton of missing species/spawns/drops/etc and what looks like completely reworked map data. Oh, and all of PrimitivePlus is gone.

Technical note: The informnation we have is that some fields have been changed from 32 to 64-bit, including export section offsets/sizes and support for 64-bit TArrays. Not sure how much effort we'll put into further support.